Patron intérprete de operaciones aritméticas

+++ src/BackEnd/Classes/Utils/DomineOp.ts
@@ -1,26 +1,26 @@
 import { Type } from './Type';
 
 export const plus: Type[][] = [
-    [Type.INT,    Type.FLOAT,  Type.DATE, Type.INT,     Type.NULL, Type.NULL],
-    [Type.FLOAT,  Type.FLOAT,  Type.DATE, Type.FLOAT,   Type.NULL, Type.NULL],
+    [Type.INT,    Type.DOUBLE, Type.DATE, Type.INT,     Type.NULL, Type.NULL],
+    [Type.DOUBLE, Type.DOUBLE, Type.DATE, Type.DOUBLE,  Type.NULL, Type.NULL],
     [Type.DATE,   Type.DATE,   Type.NULL, Type.DATE,    Type.NULL, Type.NULL],
-    [Type.INT,    Type.FLOAT,  Type.DATE, Type.VARCHAR, Type.NULL, Type.NULL],
+    [Type.INT,    Type.DOUBLE, Type.DATE, Type.VARCHAR, Type.NULL, Type.NULL],
     [Type.NULL,   Type.NULL,   Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
     [Type.NULL,   Type.NULL,   Type.NULL, Type.NULL,    Type.NULL, Type.NULL]
 ]
 
 export const minus: Type[][] = [
-    [Type.INT,    Type.FLOAT,  Type.DATE, Type.INT,     Type.NULL, Type.NULL],
-    [Type.FLOAT,  Type.FLOAT,  Type.DATE, Type.FLOAT,   Type.NULL, Type.NULL],
+    [Type.INT,    Type.DOUBLE, Type.DATE, Type.INT,     Type.NULL, Type.NULL],
+    [Type.DOUBLE, Type.DOUBLE, Type.DATE, Type.DOUBLE,  Type.NULL, Type.NULL],
     [Type.DATE,   Type.DATE,   Type.NULL, Type.DATE,    Type.NULL, Type.NULL],
-    [Type.INT,    Type.FLOAT,  Type.DATE, Type.VARCHAR, Type.NULL, Type.NULL],
+    [Type.INT,    Type.DOUBLE, Type.DATE, Type.VARCHAR, Type.NULL, Type.NULL],
     [Type.NULL,   Type.NULL,   Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
     [Type.NULL,   Type.NULL,   Type.NULL, Type.NULL,    Type.NULL, Type.NULL]
 ]
 
 export const mult: Type[][] = [
-    [Type.INT,    Type.FLOAT,  Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
-    [Type.FLOAT,  Type.FLOAT,  Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
+    [Type.INT,    Type.DOUBLE, Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
+    [Type.DOUBLE, Type.DOUBLE, Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
     [Type.NULL,   Type.NULL,   Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
     [Type.NULL,   Type.NULL,   Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
     [Type.NULL,   Type.NULL,   Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
@@ -28,8 +28,8 @@ export const mult: Type[][] = [
 ]
 
 export const div: Type[][] = [
-    [Type.INT,    Type.FLOAT,  Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
-    [Type.FLOAT,  Type.FLOAT,  Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
+    [Type.INT,    Type.DOUBLE, Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
+    [Type.DOUBLE, Type.DOUBLE, Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
     [Type.NULL,   Type.NULL,   Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
     [Type.NULL,   Type.NULL,   Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
     [Type.NULL,   Type.NULL,   Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
@@ -37,8 +37,8 @@ export const div: Type[][] = [
 ]
 
 export const mod: Type[][] = [
-    [Type.INT,    Type.FLOAT,  Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
-    [Type.FLOAT,  Type.FLOAT,  Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
+    [Type.INT,    Type.DOUBLE, Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
+    [Type.DOUBLE, Type.DOUBLE, Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
     [Type.NULL,   Type.NULL,   Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
     [Type.NULL,   Type.NULL,   Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
     [Type.NULL,   Type.NULL,   Type.NULL, Type.NULL,    Type.NULL, Type.NULL],
